+++ App/app/src/main/java/com/example/geoagenda/MainActivity.kt
@@ -35,7 +35,7 @@ class MainActivity : AppCompatActivity() {
         // Passing each menu ID as a set of Ids because each
         // menu should be considered as top level destinations.
         appBarConfiguration = AppBarConfiguration(setOf(
-                R.id.nav_home, R.id.nav_gallery, R.id.nav_slideshow, R.id.nav_logOut, R.id.nav_settings), drawerLayout)
+                R.id.nav_rec, R.id.nav_categories, R.id.nav_locations, R.id.nav_logOut), drawerLayout)
         setupActionBarWithNavController(navController, appBarConfiguration)
         navView.setupWithNavController(navController)
     }
