+++ App/app/src/main/java/com/example/geoagenda/ui/categories/CategoriesFragment.kt
@@ -1,0 +1,25 @@
+package com.example.geoagenda.ui.categories
+
+import android.os.Bundle
+import android.view.LayoutInflater
+import android.view.View
+import android.view.ViewGroup
+import androidx.fragment.app.Fragment
+import androidx.lifecycle.ViewModelProviders
+import com.example.geoagenda.R
+
+class CategoriesFragment : Fragment() {
+
+    private lateinit var categoriesViewModel: CategoriesViewModel
+
+    override fun onCreateView(
+            inflater: LayoutInflater,
+            container: ViewGroup?,
+            savedInstanceState: Bundle?
+    ): View? {
+        categoriesViewModel =
+                ViewModelProviders.of(this).get(CategoriesViewModel::class.java)
+        val root = inflater.inflate(R.layout.fragment_categories, container, false)
+        return root
+    }
+}
