+++ App/app/src/main/java/com/example/geoagenda/ui/categories/CategoriesViewModel.kt
@@ -1,13 +1,13 @@
-package com.example.geoagenda.ui.slideshow
+package com.example.geoagenda.ui.categories
 
 import androidx.lifecycle.LiveData
 import androidx.lifecycle.MutableLiveData
 import androidx.lifecycle.ViewModel
 
-class SlideshowViewModel : ViewModel() {
+class CategoriesViewModel : ViewModel() {
 
     private val _text = MutableLiveData<String>().apply {
-        value = "This is slideshow Fragment"
+        value = "This is gallery Fragment"
     }
     val text: LiveData<String> = _text
 }
