+++ App/app/src/main/java/com/example/geoagenda/ui/locations/LocationsViewModel.kt
@@ -1,13 +1,13 @@
-package com.example.geoagenda.ui.gallery
+package com.example.geoagenda.ui.locations
 
 import androidx.lifecycle.LiveData
 import androidx.lifecycle.MutableLiveData
 import androidx.lifecycle.ViewModel
 
-class GalleryViewModel : ViewModel() {
+class LocationsViewModel : ViewModel() {
 
     private val _text = MutableLiveData<String>().apply {
-        value = "This is gallery Fragment"
+        value = "This is slideshow Fragment"
     }
     val text: LiveData<String> = _text
 }
